Возвращает отображение маршрута, дат и стоимости в приложение
