+++ src/main.js
@@ -4,8 +4,8 @@ import FilterController from './controllers/filter.js';
 import LoadingComponent from './components/loading.js';
 import PageHeaderContainerComponent from './components/page-header/page-header-container.js';
 import PageNavigationComponent from './components/page-header/page-navigation.js';
-// import TripCostComponent from './components/page-header/trip-cost.js';
-// import TripRouteComponent from './components/page-header/trip-route.js';
+import TripCostComponent from './components/page-header/trip-cost.js';
+import TripRouteComponent from './components/page-header/trip-route.js';
 import TripEventsModel from './models/trip-events.js';
 import TripEventsBoardComponent from './components/page-main/trip-events/trip-event-board.js';
 import TripEventsBoardController from './controllers/trip-events-board.js';
@@ -19,7 +19,6 @@ const SERVER_URL = `https://11.ecmascript.pages.academy/big-trip/points/`;
 const tripMain = document.querySelector(`.trip-main`);
 const pageBodyContainer = document.querySelector(`main .page-body__container`);
 const tripControls = tripMain.querySelector(`.trip-controls`);
-// const tripInfoContainer = tripMain.querySelector(`.trip-info`);
 const firstTitle = tripControls.querySelector(`h2`);
 
 const api = new API(AUTHORIZATION, SERVER_URL);
@@ -32,8 +31,8 @@ const buttonAddNewEventComponent = new ButtonAddNewEventComponent(tripEventsMode
 const tripEventsBoardComponent = new TripEventsBoardComponent();
 const tripEventsBoardController = new TripEventsBoardController(tripEventsBoardComponent, tripEventsModel);
 const tripStatisticsComponent = new TripStatisticsComponent(tripEventsModel);
-// const tripRouteComponent = new TripRouteComponent(tripEventsModel);
-// const tripCostComponent = new TripCostComponent(tripEventsModel);
+const tripRouteComponent = new TripRouteComponent(tripEventsModel);
+const tripCostComponent = new TripCostComponent(tripEventsModel);
 
 render(tripMain, pageHeaderContainerComponent, RenderPosition.AFTERBEGIN);
 render(firstTitle, pageNavigationComponent, RenderPosition.AFTEREND);
@@ -41,12 +40,6 @@ render(tripMain, buttonAddNewEventComponent);
 render(pageBodyContainer, loadingComponent);
 render(pageBodyContainer, tripEventsBoardComponent);
 render(pageBodyContainer, tripStatisticsComponent);
-
-// if (tripEventsObjects.length > 0) {
-//   render(tripInfoContainer, tripRouteComponent);
-// }
-
-// render(tripInfoContainer, tripCostComponent);
 
 filterController.render();
 buttonAddNewEventComponent.setClickHandler();
@@ -71,5 +64,10 @@ api.getTripEvents()
   .then((tripEvents) => {
     tripEventsModel.setTripEvents(tripEvents);
     remove(loadingComponent);
+
+    const tripInfoContainer = tripMain.querySelector(`.trip-info`);
+
+    render(tripInfoContainer, tripRouteComponent);
+    render(tripInfoContainer, tripCostComponent);
     tripEventsBoardController.render();
   });
